Add GStreamerProcess proof-of-concept
Each GStreamerProcess has a multiprocessing.Process subclass instance.
This Process is used to run a GStreamer pipeline.  It accepts the
following commands through a pipe connection:

    Command
    -------
    'start' -> prompts user for video source and starts the video
            pipeline
    'stop' -> stops the pipeline
    'finish' -> closes the child process

as a request dictionary form:
    {'command': <command>}

Note that if `'ack': True` is added to any request, a response will be
sent back over the pipe upon handling of the request by the child
process.